fix 0.0.10.6: refined removeTeachers(Long id, List<Long> ids) in CourseServiceImpl and moved mappers to their own package
