+++ src/main/java/com/senla/courses/service/students/StudentServiceImpl.java
@@ -3,9 +3,9 @@ package com.senla.courses.service.students;
 import com.senla.courses.dao.StudentDAO;
 import com.senla.courses.dao.UserDAO;
 import com.senla.courses.dto.StudentDTO;
-import com.senla.courses.dto.StudentMapper;
+import com.senla.courses.mapper.StudentMapper;
 import com.senla.courses.dto.UserDTO;
-import com.senla.courses.dto.UserMapper;
+import com.senla.courses.mapper.UserMapper;
 import com.senla.courses.exception.EmptyListException;
 import com.senla.courses.exception.NotFoundException;
 import com.senla.courses.model.Student;
